% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_centrality.R
\name{assess_centrality}
\alias{assess_centrality}
\title{assess centrality measure to the top network}
\usage{
assess_centrality(subnetwork, graph)
}
\arguments{
\item{subnetwork}{portion of graph, top network}

\item{graph}{igraph object, total graph}
}
\value{
dataframe with degree, betwenness, closeness and
eigenvector centrality measures
}
\description{
Function for assess centrality measures to the top network
}
\examples{
\dontrun{assess_centrality(subnetwork, graph)}
}
