% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_modularity.R
\name{assess_modularity}
\alias{assess_modularity}
\title{Assess modularity}
\usage{
assess_modularity(
  G = NULL,
  vertices = NULL,
  ranks = NULL,
  commMethods = c("fastgreedy", "multilev"),
  BPPARAM = NULL
)
}
\description{
Assess modularity
}
