% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adjND.R
\name{calc_adjND}
\alias{calc_adjND}
\title{Calculation of permutation-adjusted network smoothing index}
\usage{
calc_adjND(X0, W, eps, k=99, mode=c("S", "Xs"), returnPerm=FALSE, ...)
}
\arguments{
\item{X0}{input matrix}

\item{W}{symmetrically normalized adjacency matrix W = D^-1 A D^-1, see normalize_adj_mat function}

\item{eps}{numeric value}

\item{k}{number of permutations}

\item{mode}{whether to use network smoothing index (S) (default) or
network smoothing values (Xs)}

\item{BPPARAM}{An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using SerialParam(). See ?bplapply.}

\item{returnPerm}{return permutations}

\item{...}{additional parameters of ND}
}
\value{
\code{data.frame} with X0, Xs, S, p and Sp
}
\description{
Calculation of permutation-adjusted network smoothing index
}
\examples{
\dontrun{calc_adjND(X0, W, eps, k=99, mode=c("S", "Xs"), returnPerm=FALSE)}
}
