% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dS.R
\name{calc_dS}
\alias{calc_dS}
\title{Calculation of delta network smoothing indexes}
\usage{
calc_dS(X0 = NULL, W = NULL, classes = NULL, eps = NULL, ...)
}
\arguments{
\item{X0}{input matrix}

\item{W}{normalized adjancency matrix}

\item{classes}{numeric vector of \{1,2\}}

\item{eps}{1-row matrix with two columns for the eps values of class 1 and 2
respectively}

\item{...}{additional parameteres of NPATools::ND()}
}
\description{
Calculation of delta network smoothing indexes
}
