% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_omega_i.R
\name{calc_omega_i}
\alias{calc_omega_i}
\title{Calculation of omega i (internal)}
\usage{
calc_omega_i(idx = NULL, Ai = NULL, dSprod = NULL)
}
\arguments{
\item{idx}{index}

\item{Ai}{matrix}

\item{dSprod}{matrix with products of delta S}
}
\value{
n
}
\description{
Calculation of omega i (internal)
}
