% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_X0_mean.R
\name{calc_X0_mean}
\alias{calc_X0_mean}
\title{Calculate X0 mean}
\usage{
calc_X0_mean(X0 = NULL, classes = NULL)
}
\arguments{
\item{X0}{vector or matrix composed of column vectors with initial 
distribution of information}

\item{classes}{vector of class labels for the columns of X0. Values must be
equal to 1 or 2}
}
\description{
Calculate X0 mean
}
