% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.R
\name{CCF}
\alias{CCF}
\title{Connected component fraction}
\usage{
CCF(x = NULL, min_size = 2)
}
\arguments{
\item{x}{igraph object}

\item{min_size}{min size of a connected component}
}
\description{
Connected component fraction
}
