% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_perm.R
\name{classes_perm}
\alias{classes_perm}
\title{Generate N permutations of a binary vector}
\usage{
classes_perm(z,N)
}
\arguments{
\item{z}{binary vector with values in \{1, 2\}}

\item{N}{number of permutations}
}
\value{
list of N permutations
}
\description{
Generate N permutations of a binary vector
}
\examples{
classes_perm(z,N)
}
