% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_partc_modulez.R
\name{cmp_pc_wmz}
\alias{cmp_pc_wmz}
\title{cmp_pc_wmz Comparison between functional cartographies}
\usage{
cmp_pc_wmz(PcWmz_A, PcWmz_B)
}
\arguments{
\item{PcWmz_A}{dataframe 1 resulting from pc_wmz}

\item{PcWmz_B}{dataframe 2 resulting from pc_wmz}
}
\value{
dataframe with deltaP, deltaZ, changeRegion, euclideanDistance, 
pc and wm_z from first and second network
}
\description{
Comparison between functional cartographies
}
\examples{
\dontrun{cmp_pc_wmz(PcWmz_A, PcWmz_B)}
}
