% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp_top_net_scores.R
\name{cmp_top_net_scores}
\alias{cmp_top_net_scores}
\title{Compare top networks scores}
\usage{
cmp_pc_wmz(PcWmz_A, PcWmz_B)
}
\arguments{
\item{ae_res_full}{dataframe resulting from 'assess_enrichment'}

\item{resSp}{dataframe resulting from resSp}

\item{start}{start}

\item{end}{end}

\item{by}{by}

\item{threshold}{threshold}
}
\value{
dataframe with deltaP, deltaZ, changeRegion, euclideanDistance,
pc and wm_z from first and second network
}
\description{
Define significance for enriched genes
}
\examples{
\dontrun{cmp_pc_wmz(PcWmz_A, PcWmz_B)}
}
