% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_net.R
\name{comm_net}
\alias{comm_net}
\title{reduce the network as a community network}
\usage{
comm_net(g)
}
\arguments{
\item{g}{top network, with vertex attribute "comm_id"}
}
\value{
dataframe with number of shared links between couple of communities
}
\description{
Function for reduce the network as a community network
}
\examples{
\dontrun{comm_net(g)}
}
