% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_X0_ND.R
\name{compare_X0_ND}
\alias{compare_X0_ND}
\title{compare_X0_ND}
\usage{
compare_X0_ND(
  G = NULL,
  X0_ranked_by_ND = NULL,
  X0_ranked_by_X0 = NULL,
  norm = FALSE,
  do.plot = FALSE,
  file = "./compare_X0_ND.jpg",
  ...
)
}
\description{
compare_X0_ND
}
