% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Assessment of the interactome coverage of the input scores}
\usage{
coverage(g, omics, outFile="coverage.jpg")
}
\arguments{
\item{g}{interactome, igraph object}

\item{omic}{input score data on which we calculate the coverage,
need to have a column names "scores" (typically a p-value) and a
column named "gene_id" to match with the graph attribute "name"}

\item{outFile}{name of the output file}
}
\description{
Assessment of the interactome coverage of the input scores
}
\examples{
\dontrun{coverage(g, omics, outFile="coverage.jpg")}
}
