% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmfind.R
\docType{package}
\name{dmfind}
\alias{dmfind}
\title{dmfind: Network diffusion-based analysis of omics for the identifcation of differentially enriched modules.}
\description{
The dmfind package provides three categories of important functions:
dmfind, bar and baz.
}
\section{Foo functions}{

The dmfind functions ...
}

