% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmfind_demo.R
\docType{data}
\name{dmfind_demo}
\alias{dmfind_demo}
\title{Demo data of dmfind}
\format{
An object of class \code{list} of length 15.
}
\usage{
dmfind_demo
}
\description{
Demo data of dmfind
}
\keyword{datasets}
