% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MCp.R
\name{.calc_MCp}
\alias{.calc_MCp}
\title{calc_Mcp}
\usage{
.calc_MCp(perms, real)
}
\arguments{
\item{perms}{list of permuted statistics}

\item{real}{vector of real statistics}
}
\value{
n
}
\description{
Help function
}
