% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_weights.R
\name{edge_weights}
\alias{edge_weights}
\title{Function for the assignment of edges weights in order to better visualize the communities}
\usage{
edge_weights(community, network, weightWithin=100, weightBetween=1)
}
\arguments{
\item{community}{list with the mandatory field "membership"}

\item{network}{pathway network}

\item{weightWithin}{value to weight the attraction between two vertices of the same community}

\item{weightBetween}{value to weight the attraction between two vertices of two disntict communities}
}
\value{
vector of edge weights
}
\description{
Function for the assignment of edges weights in order to better visualize the communities
}
\examples{
edge_weights(community, network, weightWithin=100, weightBetween=1)
}
