% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFDR.R
\name{eFDR}
\alias{eFDR}
\title{empirical False Discovery Rate}
\usage{
eFDR(realValues, allValues, BPPARAM = NULL)
}
\arguments{
\item{realValues}{vactor of all values (real + permuted) values}

\item{mc.cores}{number of cores}
}
\description{
empirical False Discovery Rate
}
