% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_significance.R
\name{enrichment_significance}
\alias{enrichment_significance}
\title{enrichment_significance Define significance for enriched genes}
\usage{
cmp_pc_wmz(PcWmz_A, PcWmz_B)
}
\arguments{
\item{ae_res_full}{dataframe resulting from 'assess_enrichment'}

\item{mNDp}{dataframe resulting from mNDp}

\item{start}{start}

\item{end}{end}

\item{by}{by}

\item{threshold}{threshold}
}
\value{
dataframe with deltaP, deltaZ, changeRegion, euclideanDistance,
pc and wm_z from first and second network
}
\description{
Define significance for enriched genes
}
\examples{
\dontrun{cmp_pc_wmz(PcWmz_A, PcWmz_B)}
}
