% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{es}
\alias{es}
\title{Enrichment Score}
\usage{
es(idx, x, le=FALSE)
}
\arguments{
\item{idx}{vector of indices a subset of elements of x}

\item{x}{named vector, ranked list}

\item{le}{TRUE/FALSE}
}
\value{
enrichment score
}
\description{
function to calculate the enrichment score
}
\examples{
\dontrun{es(idx, x, le=FALSE)}
}
