% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_eps.R
\name{eval_eps}
\alias{eval_eps}
\title{Evaluate the effect of epsilon on NSI}
\usage{
eval_eps(X0, Xs, eps = NULL, G, top = min(300, nrow(X0)))
}
\arguments{
\item{X0}{initial condition matrix}

\item{Xs}{network smoothing result}

\item{eps}{matrix with epsilon values (rows) for each column of X0, Xs}

\item{G}{matrix with epsilon values (rows) for each column of X0, Xs}

\item{top}{matrix with epsilon values (rows) for each column of X0, Xs}
}
\value{
list with sn_g and sn_summary elements
}
\description{
Evaluate the effect of epsilon on NSI
}
