% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sign_conn_comp.R
\name{find_sign_conn_comp}
\alias{find_sign_conn_comp}
\title{Finding significant connected components}
\usage{
find_sign_conn_comp(NR_summary, NR_p_thr = 0.05, min_rank = 50, max_rank = 300)
}
\arguments{
\item{NR_summary}{data.frame with NR_summary}

\item{NR_p_thr}{threshold for NR p value}

\item{min_rank}{minimum rank}

\item{max_rank}{maximum rank}
}
\description{
Finding significant connected components
}
