% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_cartography.R
\name{functional_cartography}
\alias{functional_cartography}
\title{pc_wmz Calculate participation coefficient and within module degree z-score}
\usage{
pc_wmz(topNetwork, allNames = FALSE, extremePoints = FALSE,
chooseYourGenes = NULL, outFile="pc_wmz.jpg", doPlot=TRUE)
}
\arguments{
\item{topNetwork}{network with some parameters: "comm_id" for community
membership and "label" for labels (for plotting purpose).}

\item{labelBy}{vertex attribute of topNetwork to use for labels}

\item{outFile}{filename for saving}
}
\value{
dataframe with pc, wmz, region, description and role
}
\description{
Calculate participation coefficient
and within module degree z-score
}
\examples{
\dontrun{pc_wmz(topNetwork, allNames = FALSE, extremePoints = FALSE,
chooseYourGenes = NULL, outFile="pc_wmz.jpg", doPlot=TRUE}
}
