% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g400esm.R
\docType{data}
\name{g400esm}
\alias{g400esm}
\title{An example subnetwork extracted from STRING}
\format{
igraph object
}
\source{
\url{https://string-db.org}
}
\usage{
data(g400esm)
}
\description{
An example subnetwork extracted from STRING
}
\keyword{datasets}
