% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.max.conn.comp.R
\name{get.max.conn.comp}
\alias{get.max.conn.comp}
\title{Extract the maximum connected components}
\usage{
get.max.conn.comp(x, verbose = TRUE)
}
\arguments{
\item{x}{network}

\item{verbose}{TRUE/FALSE}
}
\description{
Extract the maximum connected components
}
