% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.n.conn.comp.R
\name{get.n.conn.comp}
\alias{get.n.conn.comp}
\title{Extract a connected components of size n}
\usage{
get.n.conn.comp(x, n)
}
\arguments{
\item{x}{network}

\item{n}{size of connected component}

\item{verbose}{TRUE/FALSE}
}
\description{
Extract the maximum connected components
}
