% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_conn_comp.R
\name{get_max_conn_comp}
\alias{get_max_conn_comp}
\title{Extract the maximum connected components}
\usage{
get_max_conn_comp(x, verbose=TRUE)
}
\arguments{
\item{x}{network}

\item{verbose}{TRUE/FALSE}
}
\value{
maximum connected component
}
\description{
Extract the maximum connected components
}
\examples{
get_max_conn_comp(x, verbose=TRUE)
}
