% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_module_label.R
\name{get_module_label}
\alias{get_module_label}
\title{Get module label}
\usage{
get_module_label(xy, f, labs)
}
\arguments{
\item{xy}{layout coordinates}

\item{f}{factor describing modules}

\item{labs}{name vector of labels. Names  must be equal to unique values of f.}
}
\value{
data
}
\description{
Get module label
}
\examples{
get_module_label(xy, f, labs)
}
