% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_conn_comp.R
\name{get_n_conn_comp}
\alias{get_n_conn_comp}
\title{Extract a connected components of size n}
\usage{
get_n_conn_comp(x,n)
}
\arguments{
\item{x}{network}

\item{n}{size of connected component}

\item{verbose}{TRUE/FALSE}
}
\value{
subgraph
}
\description{
Extract the maximum connected components
}
\examples{
get_n_conn_comp(x,n)
}
