% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nconn_comp.R
\name{get_nconn_comp}
\alias{get_nconn_comp}
\title{Extract a connected components of size n}
\usage{
get_nconn_comp(x,n)
}
\arguments{
\item{x}{network}

\item{n}{size of connected component}
}
\value{
subgraph
}
\description{
Extract the maximum connected components
}
\examples{
\dontrun{get_nconn_comp(x,n)}
}
