% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_map.R
\name{linear_map}
\alias{linear_map}
\title{Linear map}
\usage{
linear_map(x, yMin, yMax)
}
\arguments{
\item{x}{numeric vector}

\item{yMin}{min value of the new distribution}

\item{yMax}{max value of the new distribution}
}
\value{
linear map y
}
\description{
Linear map
}
\examples{
\dontrun{linear_map(x, yMin, yMax)}
}
