% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_comparison.R
\name{NetComp}
\alias{NetComp}
\title{Network comparison}
\usage{
NetComp(NetList, union=FALSE, intersection=FALSE,
centrMeas=FALSE, ji=FALSE)
}
\arguments{
\item{NetList}{interactomes, list of igraph objects}

\item{union}{Perform the union of the input set of networks}

\item{intersection}{Perform the insersection of the input set of networks}

\item{centrMeas}{Calculate the centrality measures for every input network}

\item{ji}{Calculation of the Jaccard index}

\item{oc}{Calculation of the overlap Coefficient}
}
\value{
A list containing the occurrences matrix and, if selected, the union of the networks,
the intersection of the networks, a dataframe containing the centrality
measure for every network and the Jaccard Index.
}
\description{
Network comparison
}
\examples{
\dontrun{NetComp(NetList, union=FALSE, intersection=FALSE,
centrMeas=FALSE, ji=FALSE)}
}
