% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR.R
\name{NR}
\alias{NR}
\title{Local Monte Carlo p values}
\usage{
NR(G, ranked_vector, k = 99, mc.cores = 1, null_model = "A", norm = "n")
}
\arguments{
\item{G}{gene x gene undirected interaction graph}

\item{ranked_vector}{scores vector; it must have the same names and size of the vertices of G}

\item{k}{number of permutations}

\item{mc.cores}{number of cores}

\item{null_model}{null model type}
}
\description{
Local Monte Carlo p values
}
