% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsi.R
\name{nsi}
\alias{nsi}
\title{Network Smoothing Index}
\usage{
nsi(X0 = NULL, Xs = NULL, eps = NULL)
}
\arguments{
\item{X0}{input matrix/vector}

\item{Xs}{smoothed matrix/vector}

\item{eps}{numeric value}
}
\value{
network smoothing index S
}
\description{
Network Smoothing Index
}
