% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_perm.R
\name{omega_perm}
\alias{omega_perm}
\title{Calculation of permuted omega function}
\usage{
omega_perm(idx = NULL, dS = NULL, Ai = NULL)
}
\arguments{
\item{idx}{vector of randomly resampled dS scores;}

\item{dS}{scores vector; it must have the same names
and size of the vertices of G}

\item{Ai}{adjacency matrix relative to the elements of dS
the names must be the same as dS's}
}
\value{
vector of permuted omega
}
\description{
Calculation of permuted omega function
}
