% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partc_modulez.R
\name{pc_wmz}
\alias{pc_wmz}
\title{pc_wmz Calculate participation coeffcient and within module degree z-score}
\usage{
pc_wmz(
  top_network,
  all_names = FALSE,
  extreme_points = FALSE,
  choose_your_genes = NULL,
  out_file = "pc_wmz.jpg",
  do_plot = TRUE
)
}
\arguments{
\item{top_network}{to do}

\item{all_names}{to do}
}
\description{
pc_wmz Calculate participation coeffcient and within module degree z-score
}
