% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fc.R
\name{plot_fc}
\alias{plot_fc}
\title{Plot functional cartography}
\usage{
plot_fc(
  topNetworks = NULL,
  labelBy = NULL,
  outFile = NULL,
  hCut = NULL,
  useP = FALSE
)
}
\description{
Plot functional cartography
}
