% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Extraction and plotting of connected subnetworks from a graph 
and a set of selected vertices}
\usage{
plot_network(graph, colorBy=NULL, colorQuant=TRUE, labelBy="name", 
pal=NULL, plotOutfile="graph.jpg", community=NULL, commWin=2, commWb=1, 
lo=NULL, legendOff=FALSE, ...)
}
\arguments{
\item{graph}{igraph object}

\item{colorBy}{object to be colored by an attribute}

\item{colorQuant}{TRUE/FALSE}

\item{labelBy}{select a feature to label the graph}

\item{pal}{palette to color the graph}

\item{plotOutfile}{output file}

\item{community}{communities from network analysis}

\item{commWin}{distance of the vertices inside a community}

\item{commWb}{distance between communities}

\item{lo}{layout}

\item{legendOff}{legend}

\item{...}{additional arguments}
}
\value{
plot network
}
\description{
Extraction and plotting of connected subnetworks 
from a graph and a set of selected vertices
}
\examples{
\dontrun{if (require("igraph")) plot_network(graph, colorBy=NULL, 
colorQuant=TRUE, labelBy="name", pal=NULL, plotOutfile="graph.jpg", 
community=NULL, commWin=2, commWb=1, lo=NULL, legendOff=FALSE)}
}
