% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NR.R
\name{plot_NR}
\alias{plot_NR}
\title{Plot the results of eval_eps}
\usage{
plot_NR(nrRes=NULL, signCompTable=NULL, file="NR.jpg")
}
\arguments{
\item{nrRes}{network resampling result}

\item{signCompTable}{?}

\item{file}{filename}
}
\value{
plot
}
\description{
Plot the results of eval_eps
}
\examples{
plot_NR(nrRes=NULL, signCompTable=NULL, file="NR.jpg")
}
