% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NSI.R
\name{plot_NSI}
\alias{plot_NSI}
\title{Scatter Plot of NSI values}
\usage{
plot_NSI(nsiRes = NULL, x = "X0", y = NULL, column = 1, initSigGenes = NULL)
}
\arguments{
\item{nsiRes}{result of calc_dS() or cal_adjND()}

\item{x}{what to put on the horizontal axis}

\item{y}{what to put on the vertical axis}

\item{column}{column to consider}

\item{initSigGenes}{optional vector of initial sifnigicant genes}
}
\description{
Plot NSI
}
