% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_S.R
\name{plot_S}
\alias{plot_S}
\title{Plot S}
\usage{
plot_S(calcAdjNDres=NULL, X0, initSigGenes=NULL, file="S.jpg")
}
\arguments{
\item{calcAdjNDres}{XS Matrix from ND}

\item{X0}{X0 matrix}

\item{initSigGenes}{NULL}

\item{file}{filename}
}
\value{
plot
}
\description{
Plot S
}
\examples{
\dontrun{plot_S(calcAdjNDres=NULL, X0, initSigGenes=NULL, file="S.jpg")}
}
