% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round}
\alias{round}
\title{Scientific approximation}
\usage{
round(x,y)
}
\arguments{
\item{x}{numeric}

\item{y}{number of digits}
}
\value{
truncated output
}
\description{
scientific approximation
}
\examples{
\dontrun{round(x,y)}
}
