% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2dem.R
\name{sm2dem}
\alias{sm2dem}
\title{Pipeline for the extraction of differentially enriched modules from smoothed matrices}
\usage{
sm2dem(
  X0,
  Xs,
  G,
  classes,
  eps = 1,
  use.exp = FALSE,
  median = FALSE,
  classes_perm = NULL,
  mc.cores = 2,
  ranking_type = c("dR", "dS", "dXs", "dX0"),
  NR_p_thr = 0.05,
  NR_k = 100,
  min_module_size_final = 10,
  min_rank = 100,
  max_rank = 500,
  NR_max_rank = 600,
  min_subnet_size = 2,
  plot_flag = FALSE,
  plot_outfile = "graph.jpg",
  vector_mode = FALSE,
  ...
)
}
\arguments{
\item{X0}{input matrix}

\item{Xs}{input matrix}

\item{G}{gene x gene undirected interaction graph}

\item{classes}{numeric vector of \{1,2\}}

\item{eps}{numeric value}

\item{use.exp}{TRUE/FALSE whether to use the exponent at the denominator}

\item{median}{TRUE/FALSE whether to use median instead of mean}

\item{classes_perm}{list of numeric vectors of \{1,2\}}

\item{mc.cores}{number of cores}

\item{ranking_type}{criterion for ranking the molecular entities}

\item{NR_p_thr}{threshold for NR p value}

\item{NR_k}{number of permutations}

\item{min_module_size_final}{minimun size of final module}

\item{min_rank}{minimum rank}

\item{max_rank}{maximum rank}

\item{NR_max_rank}{network resampling will be applied from 1 to NR_max_rank}

\item{min_subnet_size}{minimum number of genes connected in the subnetworks to be extracted}

\item{plot_flag}{whether to plot or not the network}

\item{plot_outfile}{file name of the output figure}

\item{vector_mode}{TRUE/FALSE}

\item{...}{further arguments to igraph.plot}
}
\description{
Pipeline for the extraction of differentially enriched modules from smoothed matrices
}
