% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_gs_score.R
\name{std_gs_score}
\alias{std_gs_score}
\title{Calculation of pathway score matrix}
\usage{
std_gs_score(x, gsList, gsScore, useMedian=FALSE, scalingEps=1, 
mc.cores=1)
}
\arguments{
\item{x}{input matrix}

\item{gsList}{gene set list}

\item{gsScore}{gene set score}

\item{useMedian}{whether to use median (TRUE) or mean (FALSE)}

\item{scalingEps}{numeric value}

\item{mc.cores}{number of cores}
}
\value{
matrix of scaled pathway scores
}
\description{
Calculation of pathway score matrix
}
\examples{
\dontrun{std_gs_score(x, gsList, gsScore, useMedian=FALSE, scalingEps=1, 
mc.cores=1)}
}
