% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.v12.Entrez.ntm.700.400.k3.R
\docType{data}
\name{string.v12.Entrez.ntm.700.400.k3}
\alias{string.v12.Entrez.ntm.700.400.k3}
\title{Interactome derived from STRING db v12.0
igraph object with 17288 genes and 174962 interactions. Only interactions with confidence >=700 and the top 3 interactions with confidence >=400 & < 700 were considered. Confidence score was calculated without textmining, as described at the URL https://string-db.org). Original protein identifiers were mapped to Entrez Gene identifiers using mappings from NCBI Entrez (https://ftp.ncbi.nlm.nih.gov dowload date 2023-09-19) and https://string-db.org}
\format{
An object of class \code{igraph} of length 17288.
}
\usage{
string.v12.Entrez.ntm.700.400.k3
}
\description{
Interactome derived from STRING db v12.0
igraph object with 17288 genes and 174962 interactions. Only interactions with confidence >=700 and the top 3 interactions with confidence >=400 & < 700 were considered. Confidence score was calculated without textmining, as described at the URL https://string-db.org). Original protein identifiers were mapped to Entrez Gene identifiers using mappings from NCBI Entrez (https://ftp.ncbi.nlm.nih.gov dowload date 2023-09-19) and https://string-db.org
}
\keyword{datasets}
