% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv2dem.R
\name{sv2dem}
\alias{sv2dem}
\title{Pipeline for the extraction of differentially enriched modules from an input vector}
\usage{
sv2dem(
  X,
  G,
  mc.cores = 2,
  NR_p_thr = 0.05,
  NR_k = 100,
  min_module_size_final = 10,
  min_rank = 100,
  max_rank = 500,
  NR_max_rank = 600,
  min_subnet_size = 2,
  plot_flag = FALSE,
  plot_outfile = "graph.jpg",
  ...
)
}
\arguments{
\item{X}{input vector}

\item{G}{gene x gene undirected interaction graph}

\item{mc.cores}{number of cores}

\item{NR_p_thr}{threshold for NR p value}

\item{NR_k}{number of permutations}

\item{min_module_size_final}{minimun size of final module}

\item{min_rank}{minimum rank}

\item{max_rank}{maximum rank}

\item{NR_max_rank}{network resampling will be applied from 1 to NR_max_rank}

\item{min_subnet_size}{minimum number of genes connected in the subnetworks to be extracted}

\item{plot_flag}{whether to plot or not the network}

\item{plot_outfile}{file name of the output figure}

\item{...}{further arguments to igraph.plot}
}
\description{
Pipeline for the extraction of differentially enriched modules from an input vector
}
