% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_module_degree.R
\name{within_module_degree}
\alias{within_module_degree}
\title{Within-module degree}
\usage{
within_module_degree(g = NULL, memb = NULL, A = NULL, useP = FALSE)
}
\description{
Within-module degree
}
