% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_enrichment.R
\name{assess_enrichment}
\alias{assess_enrichment}
\title{assess enrichment of the top networks}
\usage{
assess_enrichment(G=NULL, topList=NULL, ranks=NULL, 
rankedVectorX0=NULL, doPlot=FALSE, critical=NULL, file="topNet_enrichment.jpg")
}
\arguments{
\item{G}{igraph object}

\item{topList}{ranked list of vertex names}

\item{ranks}{ranks of topList that will be assessed}

\item{rankedVectorX0}{ranked vector by X0}

\item{doPlot}{whether to plot or not}

\item{critical}{a possible rank to draw in the plot}

\item{file}{output file name}
}
\value{
list of values
}
\description{
Perform enrichment analysis of the top network 
based on a ranked vector
}
\examples{
\dontrun{assess_enrichment(G=NULL, topList=NULL, ranks=NULL, 
rankedVectorX0=NULL, doPlot=FALSE, critical=NULL, file="topNet_enrichment.jpg")}
}
