% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_enrichment.R
\name{assess_enrichment}
\alias{assess_enrichment}
\title{assess enrichment of the top networks}
\usage{
assess_enrichment(
  G = NULL,
  top_list = NULL,
  ranks = NULL,
  ranked_vector_X0 = NULL,
  do.plot = FALSE,
  critical = NULL,
  file = "top_net_enrichment.jpg"
)
}
\description{
assess enrichment of the top networks
}
