% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_enrichment.R
\name{assess_enrichment}
\alias{assess_enrichment}
\title{assess enrichment of the top networks}
\usage{
assess_enrichment(G, topList, ranks, rankedVectorX0, doPlot=FALSE, critical, file="outfile.jpg")
}
\arguments{
\item{G}{igraph object}

\item{topList}{ranked list of vertex names}

\item{ranks}{ranks of topList that will be assessed}

\item{rankedVectorX0}{ranked vector by X0}

\item{critical}{a possible rank to draw in the plot}

\item{file}{output file name}

\item{do.plot}{whether to plot or not}
}
\value{
list of values
}
\description{
Perform enrichment analysis of the top network based on a ranked vector
}
\examples{
assess_enrichment(G, topList, ranks, rankedVectorX0, doPlot=FALSE, critical, file="outfile.jpg")
}
