% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_enrichment.R
\name{assess_enrichment}
\alias{assess_enrichment}
\title{assess enrichment of the top networks}
\usage{
assess_enrichment(
  G = NULL,
  top_list = NULL,
  ranks = NULL,
  ranked_vector_X0 = NULL,
  do.plot = FALSE,
  critical = NULL,
  file = "top_net_enrichment.jpg"
)
}
\arguments{
\item{G}{igraph object}

\item{top_list}{ranked list of vertex names}

\item{ranks}{ranks of top_list that will be assessed}

\item{ranked_vector_X0}{ranked vector by X0}

\item{do.plot}{whether to plot or not}

\item{critical}{a possible rank to draw in the plot}

\item{file}{output file name}
}
\description{
assess enrichment of the top networks
}
