% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_enrichment.R
\name{assess_enrichment}
\alias{assess_enrichment}
\title{assess enrichment of the top networks}
\usage{
assess_enrichment(
  G = NULL,
  topList = NULL,
  ranks = NULL,
  X0Vector = NULL,
  type = c("gsea", "ora"),
  k = 99,
  critical = NULL,
  minComponentSize = 2,
  minNetSize = 10,
  minKNes = 10,
  BPPARAMGsl = NULL,
  BPPARAMK = NULL
)
}
\arguments{
\item{G}{igraph object}

\item{topList}{ranked list of vertex names that will be used to define top networks}

\item{ranks}{ranks of topList that will be assessed}

\item{X0Vector}{named numeric vector that will be tested with GSEA or ORA. In case of GSEA it will be ranked by decreasing orderg, while in the case of ORA the names of the X0 values grater than 0 will be tested for enrichment, while all X0 names will be the universe.}

\item{k}{number of permutations}
}
\description{
assess enrichment of the top networks
}
