% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_adjND.R
\name{calc_adjND}
\alias{calc_adjND}
\title{Calculation of permutation-adjusted network smoothing index}
\usage{
calc_adjND(
  X0 = NULL,
  W = NULL,
  eps = NULL,
  k = 99,
  mode = c("S", "Xs"),
  BPPARAM = NULL,
  returnPerm = FALSE,
  bin_type = "number",
  method = "simple",
  cut_par = 20,
  vertex_sets = NULL,
  ...
)
}
\arguments{
\item{X0}{input matrix}

\item{W}{symmetrically normalized adjacency matrix W = D^-1 A D^-1, see normalize_adj_mat function}

\item{eps}{numeric value}

\item{k}{number of permutations}

\item{mode}{whether to use network smoothing index (S) (default) or
network smoothing values (Xs)}

\item{BPPARAM}{An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using SerialParam(); See BiocParallel::bplapply()}

\item{returnPerm}{return permutations}

\item{bin_type}{see NPATools::perm_vertices()}

\item{method}{see NPATools::perm_vertices()}

\item{cut_par}{see NPATools::perm_vertices()}

\item{vertex_sets}{see NPATools::perm_vertices()}

\item{...}{additional arguments passed to NPATools::ND()}
}
\value{
list of data frames
}
\description{
Calculation of permutation-adjusted network smoothing index
}
