% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ccf.R
\name{calculate_ccf}
\alias{calculate_ccf}
\title{Calculate ccf}
\usage{
calculate_ccf(G, rankedVectorNames)
}
\arguments{
\item{G}{igraph object}

\item{rankedVectorNames}{ranked vector of names}
}
\value{
cff
}
\description{
Calculate ccf
}
\examples{
calculate_ccf(G, rankedVectorNames)
}
