% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_X0_ND.R
\name{compare_X0_ND}
\alias{compare_X0_ND}
\title{compare_X0_ND}
\usage{
compare_X0_ND(G, X0rankedByND, X0rankedByX0, norm, do.plot=FALSE, file="./compare_X0_ND.jpg")
}
\arguments{
\item{G}{graph}

\item{X0rankedByND}{ranked vector of ND values ordered by X0}

\item{X0rankedByX0}{ranked vector of X0 values ordered by X0}

\item{norm}{FALSE}

\item{do.plot}{FALSE}
}
\value{
Comparison plot between X0 and ND
}
\description{
Generate N permutations of a binary vector
}
\examples{
compare_X0_ND(G, X0rankedByND, X0rankedByX0, norm, do.plot=FALSE, file="./compare_X0_ND.jpg")
}
