% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_eps.R
\name{eval_eps}
\alias{eval_eps}
\title{Evaluate the effect of epsilon on NSI}
\usage{
eval_eps(X0, Xs, eps=NULL, G, top)
}
\arguments{
\item{X0}{initial condition matrix}

\item{Xs}{network smoothing result}

\item{eps}{matrix with epsilon values (rows) for each column of X0, Xs}

\item{G}{igraph object}

\item{top}{top genes to consider}
}
\value{
list with sn_g and sn_summary elements
}
\description{
Evaluate the effect of epsilon on NSI
}
\examples{
eval_eps(X0, Xs, eps=NULL, G, top)
}
