% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_eps.R
\name{eval_eps}
\alias{eval_eps}
\title{Evaluate the effect of epsilon on NSI}
\usage{
eval_eps(X0 = NULL, Xs = NULL, eps = NULL, G = NULL, top = NULL)
}
\arguments{
\item{X0}{initial condition matrix}

\item{Xs}{network smoothing result}

\item{eps}{matrix with epsilon values (rows) for each column of X0, Xs}

\item{G}{igraph object}

\item{top}{top genes to consider}
}
\value{
list of data frames
}
\description{
Evaluate the effect of epsilon on NSI
}
