% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_module.R
\name{extract_module}
\alias{extract_module}
\title{Extraction and plotting of connected subnetworks from a graph and a set of selected vertices}
\usage{
extract_module(graph, selected_vertices, X0 = NULL, min_subnet_size = 2)
}
\arguments{
\item{graph}{igraph object}

\item{selected_vertices}{subset of vertices}

\item{min_subnet_size}{minimum size of subnetwork}
}
\description{
Extraction and plotting of connected subnetworks from a graph and a set of selected vertices
}
