% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_module.R
\name{extract_module}
\alias{extract_module}
\title{Extraction and plotting of connected subnetworks from a graph and a set of selected vertices}
\usage{
extract_module(graph, selectedVertices, X0=NULL, minSubnetSize=2)
}
\arguments{
\item{graph}{igraph object}

\item{selectedVertices}{subset of vertices}

\item{minSubnetSize}{minimum size of subnetwork}
}
\value{
subnetwork
}
\description{
Extraction and plotting of connected subnetworks from a graph and a set of selected vertices
}
\examples{
extract_module(graph, selectedVertices, X0=NULL, minSubnetSize=2)
}
