% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sign_conn_comp.R
\name{find_sign_conn_comp}
\alias{find_sign_conn_comp}
\title{Finding significant connected components}
\usage{
find_sign_conn_comp(NRsummary, NRpThr=0.05, minRank=50, maxRank=300)
}
\arguments{
\item{NRsummary}{data.frame with NRsummary}

\item{NRpThr}{threshold for NR p value}

\item{minRank}{minimum rank}

\item{maxRank}{maximum rank}
}
\value{
significant connected component
}
\description{
Finding significant connected components
}
\examples{
\dontrun{find_sign_conn_comp(NRsummary, NRpThr=0.05, minRank=50, 
maxRank=300)}
}
