% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.n.conn.comp.R
\name{get.n.conn.comp}
\alias{get.n.conn.comp}
\title{Get the n-sized connected components}
\usage{
get.n.conn.comp(x = NULL, n = NULL)
}
\arguments{
\item{x}{igraph object}

\item{n}{network component minimum size}
}
\description{
Get the n-sized connected components
}
