% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nconn_comp.R
\name{get_nconn_comp}
\alias{get_nconn_comp}
\title{Extract connected components of at least n vertices}
\usage{
get_nconn_comp(x, n)
}
\arguments{
\item{x}{igraph obejct}

\item{n}{minimum size of a connected component}
}
\value{
igraph object
}
\description{
Extract connected components of at least n vertices
}
\examples{
\dontrun{get_nconn_comp(x,n)}
}
