% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_subnetworks.R
\name{get_top_subnetworks}
\alias{get_top_subnetworks}
\title{Calculation of permutation-adjusted network smoothing index}
\usage{
get_top_subnetworks(G, rankedList, idx=1:length(rankedList))
}
\arguments{
\item{G}{igraph graph object}

\item{rankedList}{ranked vector; names must match V(G)$name}

\item{idx}{vector of ranks according to which networks will be extracted}
}
\value{
list with snG and snSummary elements
}
\description{
Calculation of permutation-adjusted network smoothing index
}
\examples{
get_top_subnetworks(G, rankedList, idx=1:length(rankedList))
}
