% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_score.R
\name{gs_score}
\alias{gs_score}
\title{Calculation of gene set scores}
\usage{
gs_score(x, gsList)
}
\arguments{
\item{x}{input vector}

\item{gsList}{gene set listy gene set list}
}
\value{
numeric vector of gene set scores
}
\description{
Calculation of gene set scores
}
\examples{
gs_score(x, gsList)
}
