% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR.R
\name{NR}
\alias{NR}
\title{Local Monte Carlo p values}
\usage{
NR(G, rankedVector, k=99, mc.cores=1, nullModel="A", norm="n")
}
\arguments{
\item{G}{gene x gene undirected interaction graph}

\item{rankedVector}{scores vector; it must have the same
names and size of the vertices of G}

\item{k}{number of permutations}

\item{nullModel}{null model type}

\item{norm}{type of normalization}

\item{BPPARAM}{An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using SerialParam(). See ?bplapply.}
}
\value{
list of values
}
\description{
Local Monte Carlo p values
}
\examples{
\dontrun{
NR(G, rankedVector, k=99, mc.cores=1, nullModel="A", norm="n") }
}
