% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NR.R
\name{NR}
\alias{NR}
\title{Network Resampling}
\usage{
NR(G = NULL, rankedVector = NULL, k = 99, BPPARAM = NULL)
}
\arguments{
\item{G}{igraph object}

\item{rankedVector}{ranked named vector with vertex scores; only `names(rankedVector)` in `V(G)$name` will be considered}

\item{k}{number of permutations}

\item{BPPARAM}{An optional BiocParallelParam instance determining the parallel back-end to be used during evaluation. If NULL, parallel evaluation is disabled using `SerialParam()`. See ?bplapply.}
}
\value{
list of values
}
\description{
Network Resampling
}
