% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{omega}
\alias{omega}
\title{Calculation of omega function}
\usage{
omega(G, u)
}
\arguments{
\item{G}{gene x gene undirected interaction graph}

\item{u}{ranked named list; names have to be those in V(G)$name}
}
\description{
Calculation of omega function
}
