% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partc_modulez.R
\name{pc_wmz}
\alias{pc_wmz}
\title{pc_wmz Calculate participation coefficient and within module degree z-score}
\usage{
pc_wmz(topNetwork, allNames = FALSE, extremePoints = FALSE,
chooseYourGenes = NULL, outFile="pc_wmz.jpg", doPlot=TRUE)
}
\arguments{
\item{topNetwork}{network with some parameters: "comm_id" for community
membership and "label" for labels (for plotting purpose).}

\item{allNames}{TRUE/FALSE. Used to plot all gene names.}

\item{extremePoints}{TRUE/FALSE. Used to plot extreme points of the plot,
such as the gene with max and min participation coefficient and z-score.}

\item{chooseYourGenes}{Vector of names (matching with "label" of network).
Used to plot genes name of your choice.}

\item{outFile}{filename for saving}

\item{doPlot}{TRUE/FALSE}
}
\value{
dataframe with pc, wmz, region, description and role
}
\description{
Calculate participation coefficient
and within module degree z-score
}
\examples{
\dontrun{pc_wmz(topNetwork, allNames = FALSE, extremePoints = FALSE,
chooseYourGenes = NULL, outFile="pc_wmz.jpg", doPlot=TRUE}
}
