% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Extraction and plotting of connected subnetworks from a graph and a set of selected vertices}
\usage{
plot_network(
  graph,
  color_by = NULL,
  color_quant = TRUE,
  label_by = "name",
  pal = NULL,
  plot_outfile = "graph.jpg",
  community = NULL,
  comm_w_in = 2,
  comm_w_b = 1,
  lo = NULL,
  legend.off = FALSE,
  ...
)
}
\arguments{
\item{graph}{igraph object}

\item{plot_outfile}{output file}

\item{...}{additional arguments}

\item{selected_vertices}{subset of vertices}

\item{vertices_subset}{a subset of selected_vertices (optional)}

\item{vertices_weight}{weights (optional)}

\item{vertices_label}{labels (optional)}

\item{min_subnet_size}{minimum size of subnetwork}

\item{plot_flag}{whether to plot or not the network}
}
\description{
Extraction and plotting of connected subnetworks from a graph and a set of selected vertices
}
