% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Extraction and plotting of connected subnetworks from a graph and a set of selected vertices}
\usage{
plot_network(graph, colorBy=NULL, colorQuant=TRUE, labelBy="name", pal=NULL, plotOutfile="graph.jpg", community=NULL, commWin=2, commWb=1, lo=NULL, legendOff=FALSE)
}
\arguments{
\item{graph}{igraph object}

\item{plotOutfile}{output file}

\item{...}{additional arguments}

\item{selectedVertices}{subset of vertices}

\item{verticesSubset}{a subset of selectedVertices (optional)}

\item{verticesWeight}{weights (optional)}

\item{verticesLabel}{labels (optional)}

\item{minSubnet_Size}{minimum size of subnetwork}

\item{plotFlag}{whether to plot or not the network}
}
\value{
plot network
}
\description{
Extraction and plotting of connected subnetworks from a graph and a set of selected vertices
}
\examples{
plot_network(graph, colorBy=NULL, colorQuant=TRUE, labelBy="name", pal=NULL, plotOutfile="graph.jpg", community=NULL, commWin=2, commWb=1, lo=NULL, legendOff=FALSE)
}
