% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NR.R
\name{plot_NR}
\alias{plot_NR}
\title{Plot the results of eval_eps}
\usage{
plot_NR(nrRes = NULL, signCompTable = NULL)
}
\arguments{
\item{nrRes}{result of NR()}

\item{signCompTable}{optional result of find_sign_conn_comp()}
}
\description{
Plot the results of eval_eps
}
