% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_NSI.R
\name{plot_NSI}
\alias{plot_NSI}
\title{Scatter Plot of NSI values}
\usage{
plot_S(nsiRes=NULL, X0, initSigGenes=NULL, file="S.jpg")
}
\arguments{
\item{nsiRes}{result of calc_dS() or cal_adjND()}

\item{initSigGenes}{NULL}

\item{file}{filename}
}
\value{
plot
}
\description{
Plot S
}
\examples{
\dontrun{plot_S(nsiRes=NULL, X0, initSigGenes=NULL, file="S.jpg")}
}
