% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_omega_eps.R
\name{plot_omega_eps}
\alias{plot_omega_eps}
\title{Plot the results of eval_eps}
\usage{
plot_omega_eps(evalEpsRes = NULL, column = 1)
}
\arguments{
\item{evalEpsRes}{Epsilon evaluation}

\item{column}{which columm of evalEpsRes$csX0 and evalEpsRes$csXs will be used}
}
\description{
Plot the results of eval_eps
}
