% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_omega_eps.R
\name{plot_omega_eps}
\alias{plot_omega_eps}
\title{Plot the results of eval_eps}
\usage{
plot_omega_eps(evalEpsRes, file="eval_eps_X0_omega.jpg")
}
\arguments{
\item{evalEpsRes}{Epsilon evaluation}

\item{column}{which columm of evalEpsRes$csX0 and evalEpsRes$csXs}

\item{file}{filename}
}
\value{
plot
}
\description{
Plot the results of eval_eps
}
\examples{
\dontrun{plot_omega_eps(evalEpsRes, file="eval_eps_X0_omega.jpg")}
}
