% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_S.R
\name{plot_S}
\alias{plot_S}
\title{Plot S}
\usage{
plot_S(calcAdjNDres=NULL, X0, initSigGenes=NULL, file="S.jpg")
}
\arguments{
\item{file}{filename}

\item{evalEpsRes}{Epsilon evaluation}
}
\value{
plot
}
\description{
Plot S
}
\examples{
plot_S(calcAdjNDres=NULL, X0, initSigGenes=NULL, file="S.jpg")
}
