% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI.R
\docType{data}
\name{PPI}
\alias{PPI}
\title{Gene-gene interactions from STRING db
Data frame with high confidence (>=700) gene-gene interactions. For every gene, also the top 3 interactions in the range [400, 699] are considered. Genes are identified with Entrez Gene identifiers.}
\format{
An object of class \code{data.frame} with 205055 rows and 14 columns.
}
\usage{
PPI
}
\description{
Gene-gene interactions from STRING db
Data frame with high confidence (>=700) gene-gene interactions. For every gene, also the top 3 interactions in the range [400, 699] are considered. Genes are identified with Entrez Gene identifiers.
}
\keyword{datasets}
