% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2dem.R
\name{sm2dem}
\alias{sm2dem}
\title{Pipeline for the extraction of differentially enriched modules from smoothed matrices}
\usage{
sm2demX(0, Xs, G, classes, eps=1, useExp=FALSE, median=FALSE, classesPerm=NULL, mc.cores=2, rankingType=c("dR", "dS", "dXs", "dX0"), NRpThr=0.05, NRk=100, minModuleSizeFinal=10, minRank=100, maxRank=500, NRmaxRank=600, minSubnetSize=2, plotFlag=FALSE, plotOutfile="graph.jpg", vectorMode=FALSE)
}
\arguments{
\item{X0}{input matrix}

\item{Xs}{input matrix}

\item{G}{gene x gene undirected interaction graph}

\item{classes}{numeric vector of \{1,2\}}

\item{eps}{numeric value}

\item{useExp}{TRUE/FALSE whether to use the exponent at the denominator}

\item{median}{TRUE/FALSE whether to use median instead of mean}

\item{classesPerm}{list of numeric vectors of \{1,2\}}

\item{mc.cores}{number of cores}

\item{rankingType}{criterion for ranking the molecular entities}

\item{NRpThr}{threshold for NR p value}

\item{NRk}{number of permutations}

\item{minModuleSizeFinal}{minimun size of final module}

\item{minRank}{minimum rank}

\item{maxRank}{maximum rank}

\item{NRmaxRank}{network resampling will be applied from 1 to NRmaxRank}

\item{minSubnetSize}{minimum number of genes connected in the subnetworks to be extracted}

\item{plotFlag}{whether to plot or not the network}

\item{plotOutfile}{file name of the output figure}

\item{vectorMode}{TRUE/FALSE}

\item{...}{further arguments to igraph.plot}
}
\value{
???
}
\description{
Pipeline for the extraction of differentially enriched modules from smoothed matrices
}
\examples{
sm2demX(0, Xs, G, classes, eps=1, useExp=FALSE, median=FALSE, classesPerm=NULL, mc.cores=2, rankingType=c("dR", "dS", "dXs", "dX0"), NRpThr=0.05, NRk=100, minModuleSizeFinal=10, minRank=100, maxRank=500, NRmaxRank=600, minSubnetSize=2, plotFlag=FALSE, plotOutfile="graph.jpg", vectorMode=FALSE)
}
